AGI_FASTAPI/requirements.txt
# Core requirements
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary
alembic

# Environment variables
python-dotenv

# Data validation
pydantic

# Development tools
python-multipart
httpx
